Install only flatpak configs if flatpak and a graphical env is available
This commit adds a check to only install flatpak related configs
(for now only for Qutebrowser) when flatpak and a graphica environment
like a X server is available. Because on a server/headless machine
graphical settings are not used and therefore shouldn't be installed.

Also this commit adds a script for automatic flatpak permission setting
to allow Qutebrowser to launch external programs via 'flatpak-spawn' and
to be able to talk to external editors via the '/tmp' directory