Fix .chezmoiignore weird ignore behaviour not similiar to .gitignore
I expected the ignore value `.config/i3` to also ignore `.config/i3/theme.conf` files, but this is only the case for files in the repository, not external ones.
External files will still be placed to this path, even though it is ignored.

Apparently this is by design??

So this commit works around this strange behavior of chezmoi to also
ignore externals