Improve openbrowser by accepting arguments to specify the desired browser
This commit improves openbrowser by allowing to add arguments to it.
The first argument specifies the browser (q* = qutebroweser, f* = firefox,
c* = chromium) and the next arguments are then passed to the browser.

So now the browser to be opened can be specified and also the url or special
parameters can be given