Fix i3 custom.conf workspace assignment to outputs doesn't work
This issues was caused because the definition of the workspace
variables ($ws1..) were defined in an included config and not in the
main config, and i3 has the behavior that variables are only local to
the config file except for the global config, whose variables are shared
with all included config files.
Because the variables were defined in workspace.conf and the output
assignment was done in custom.conf the variables were not available.

This commit moves the variable assignment to the main config