Fix .bin/setxkblayout was not ignored by chezmoiignore
This commit makes sure that .bin/setxkblayout is not installed when on
an headless system. Without an graphical system the script would be
useless.

Also this commit renames the chezmoi init fonts script to align it to
the naming standard of all other chezmoi init scripts
(replace - dashes in name with _ underscores)