Fix qutebrowser edit text/form-field with external editor not working
This commit fixes that the editor could not be opened on website input
with the keybinds `<Ctrl-X><Ctrl-E>` becaue the default binding in insert
mode had `<Ctrl-X><Ctrl-E>` bound, overwritting my custom bind.

Unfortunately to this, cutting text will not work with `<Ctrl-X>` but
this shouldn't be a big deal as text cutting can still be made with
the right mouse option menu - but wait are you suggesting I shall use
a mouse?