Use i3-battery-popup from main repo and not from personal fork
The main repo has merged my fork and therefore there is no need to my
personal fork, as this will require me to maintain the fork.
Since my fork was merged back this commit will use the main repo as
download source for the script.

Also the script will be downloaded not from the master branch but from a
fixed commit. This will stop breaking the chezmoi download because of
failed checksum verification if upstream changes.
Instead I will need to update the commit to download manually but I can
track verify changes made to the script