Add external AppImages like tmux and vim as binaries
This commit adds support for including external AppImages which are then
installed at user directory ~/.bin
This allows me to have all important apps (for now tmux and vim) being
already installed, without me worrying that it might not be installed on
the system, because they are now shipped with my dotfiles.

Also this provides the benefit of having always a constant version to
work with, no version differences between machines and all scripts and
plugins should be happy.

This was mainly initiated because the Ubuntu 20.04 LTS installable
tmux v3.2a is buggy and crashes when ESC sequences are entered into
commandline accidentally.
So I started to ship tmux v3.3a with chezmoi that fixes this issue.
On the same go I also ship vim v9.0.1826 which is also nice, because
some systems still run vim version 8.*

Due to AppImages nature (almost) no dependencies are required, however
one single dependency which caused problems for me, was the reliance on
fuse v2 (not fuse v3).
But not all systems allow fuse (docker) or have a incompatible version
(like Gentoo, ships only fuse v3).
To make the experience seemless, I applied a workaround that does not
require any user intervention.
If the library `libfuse2.so` is not found, try to extract the AppImage
and run it directly. This is done with the `--appimage-extract-and-run`
parameter of the AppImages.

With this supported and unsupported systems should run these AppImages
flawlessly without any problems.

Note: as the AppImages are built on Ubuntu 20.04 LTS which ships with
glibc 2.29 and AppImages need to dynamically link to it, the minimum
glibc version of the host must also be 2.29. Versions below are unlikely
to work.

Current glibc version can be extracted with:
`$(ldconfig --print-cache | grep --max-count 1 libc.so.6 | cut -d' ' -f4)`