Add qutebrowser config for non flatpak installs
This commit adds support for qutebrowser config installs without
requiring flatpak. Instead native or snap packages can be used and the
config will be correctly installed, important is that `qutebrowser` is
in `$PATH` for the detection to work.

Also this commit fixes issues with `termfg` incorrectly splitting
parameters on space characters. This is now fixed by manually enquoting
the paramaters again when given the subshell