Move i3status volume order to begin to have conistent output
This commit moves the i3status order of the volume output to the
front (beginning) of the bar. This is needed because on my laptop
the audio device name is 60+ characters long and all other information
have been pushed out to the left out of the screen.

This fixes it by moving the volume to the far left preventing it from
shifting other infos away.
Now the volume shifts itself out, the audio device name is cut, but that
is no problem as its irrelevant anyways