Allow target arguments to pwrsuspend and pwrshutdown scripts
This commit allows the user to specify the desired target for the
pwrshutdown and the pwrsuspend scripts.
Also pwrshutdown gained the reboot action.

`pwrshutdown` supports two modes:
- shutdown (default) (param: shutdown)
- reboot             (param: reboot)

`pwrsuspend` supports now three modes:
- hibernate only     (param: hibernate)
- suspend only       (param: suspend)
- try hibernate, then suspend (default)

The previous behavior is unchanged when invoked without arguments.
(See default above)

Additionally the scripts will now log their output to `~/.var/log/`
when the directory exists, if not the log is written to the console.

Also pwrshutdown has now dbus commands it can try to shutdown or reboot
the machine. This is only useable for DE with dbus support. However that
should provide passwordless shutdown/reboot when allowed.

The dbus commands are now tried first in both pwrshutdown and pwrsuspend,
as they mimick the action of DE's shutdown dialog most closest