Make .xsession executable and add to xsession's PATH the users .bin folder
This commit makes the .xsession executable for testing and
debugging it manually.

Also in the .xsession add to PATH the user ~/.bin directory so that all
applications started from i3 can access the scripts/programs found in
~/.bin
This is needed because when i3 starts no shells rc files are read.
So I needed a way to set the PATH anyways.

Note: this is only restricted to environments where the .xsession is
sources/executed like session run by Xrdp. For all other i3 startups
these changes will no apply and must rely on shell setting the PATH
variable correctly