Create .var/log with chezmoi natively not with run_after_ script
This commit removes the unnecessary run_after_ script in .bin that
created the .var/log directory.
Instead create this directory with chezmoi by adding it as a directory
to be managed by chezmoi. Now chezmoi will try to create it if not
available, while reducing the number of scripts it has to run