Move .bin folder to more common .local/bin
This commit moves the .bin folder to the more common .local/bin folder.
This is necessary so shells can detect the folder and add it automatically
to the PATH env variable. After looking into various shell initialization
scripts like ~/.bash_rc, ~/.bash_profile and others I came to the
conclusion that ~/.local/bin is a path of expected bin to be placed.
If it exist it will be added to the PATH by default on most shells.

The other consideration was to use the ~/bin path, but that seems to be
outdated and more and more apps use the ~/.local/bin prefix.
Also this would clutter my home dir with other things.

Additionally the ~/.local path is meant to mirror the /usr/local/ path,
but for the user locally.
So when compiling and installing a program from source I could install
it into ~/.local with ./configure --prefix=${HOME}/.local/
Thus I follow this convention by moving the binaries here also to that
directory