Only install AppImages on supported archs (amd64 for now)
This commit ensures that only AppImages are installed which are compatible
with the current arch.

AppImages built for amd64 won't run on arm, so ensure that only binaries
are installed which can run on target system.
Currently only amd64 is supported, but this can change should I need
those AppImages on machines with that archs.

Most AppImages are built only for amd64, but should I have enough
interest to have a working AppImages for example on Raspberry Pi
or an Android system, I might create those AppImages for those
archs myself