Run AppImages via wrapper with symlinks
This commit removes the unnecessary wrapper script of the commands that
are to run as AppImages.

For example: ~/.local/bin/vim was a wrapper script that calls
~/.local/bin/.AppImageRunner which then runs
~/.local/bin/AppImages/vim.AppImage
Since this is a long chain, I improved it and now symlinks can be used
to specify what AppImage the .AppImageRunner wrapper runs.
So the symlink ~/.local/bin/vim points to ~/.local/bin/.AppImageRunner
and AppImageRunner will understand as what it was called and runs the
specified AppImage.

This reduces a layer of wrapper script and could potentially increase
slighty the startup performance