Fix invalid ignore of tmux and vim AppImages on other archs than amd64
This commit fixes a type in the .chezmoiignore file which had an invalid
ignore path for the two AppImage symlinks.
Without this fix the symlinks would've been placed nevertheless the arch
and would break the usage of tmux and vim on those machines.
Now the symlinks are only placed on amd64 architecture, where the
AppImage is available and supported