Add first iteration of shell configuration (mainly bash)
This commit adds .bashrc and .profile to the dotfiles.
These are pretty standard configs taken from Ubuntu and source
~/.local/bin
In the future I might adjust them to my needs.

Nice thing is that .profile is read by almost all shells (except for fish
and C-shells). It contains the adding of the user ~/.local/bin to the PATH
so those binaries should be available to almost all shells.

Other shells and a personal customization are the next steps todo.

Also chezmoi will remove .bash_profile and .bash_login on every run.
This is needed to force bash to read the .profile file.