Increase bash history length and write the execute timestamp
This commit increases/doubles the history length/retention of bash
command executions. This allows for more commands to be tracked which
have been executed. Useful to find a special command for a rarely used
occasion.

Also in bash history the execution timestamp will also be saved, so its
now easier to find out when a command has been executed. Useful to track
back sessions