Fix .bashrc evaluating lesspipe could abort config parsing
This commit fixes the correct behavior on Ubuntu not working on Gentoo.
As lesspipe is implemented as some sort of script in Ubuntu the initial
.bashrc was sourcing the output of `lesspipe` but on Gentoo lesspipe is
behaving differently and thus breaks reading of the .bashrc as the
output of lesspipe cannot be parsed correctly.

As I don't even know what lesspipe is, I removed those lines and the
issue has disappeared in Gentoo. Change should not affect other
distributions, only when using lesspipe, but what exactly is that?