Fix PATH not set inside Docker container for bash
This commit fixes the issue that inside Docker containers the local
users bin directories are not added to the PATH variable.
This is due to the fact that bash inside Docker is not invoked as a
login shell and only reads .bashrc
Normally one would expect in this case that bash would inherit the
correct PATH from the parent process, but since the first process inside
a container has no parent where PATH could be inherited correctly,
PATH inside container didn't container ~/.local/bin and ~/bin

This fix does the PATH evaluation inside .bashrc aswell, to ensure PATH
is setup correctly, if not add it.