Add .chezmoiversion to ensure supported chezmoi is running
This commit makes sure that only chezmoi versions which understand
all chezmoi configurations are run.
This means atleast v2.38.0 is required to install my dotfiles.

This should cause no problems as in all cases I will run chezmoi via the
init oneliner and not install it from the packagemanager, nevertheless I
added this here.

Chezmoi v2.38.0 is the initial version with which this repo was started
with, so it is fully compatibel