Improve battery detection, especially for VMs
This commit improves the battery detection of chezmoi.
This was broken for VMs running under VMWare as they were detected to
cantain a battery but that wasn't true.
It happened because VMWare advertises the VM a charger devices,
but no battery devices.
The previous detection assumed all devices that have to to with
power management are batteries and thus the charger devices was
falsely detected as a battery.

Now only batteries are detected, and charging devices are excluded