Bump Vim AppImage version to v9.0.1876
This commit bumps the used Vim AppImage to v9.0.1876
It includes a fix which I reported to the maintainer:

When run without fuse, that is with `--appimage-extract-and-run`
instead, relative paths in arguments to Vim are broken, because the CWD
(current working directory) is not correctly set. The CWD was in the
AppImage extraction dir and not as expected the location from where the
command was invoked.

This is now fixed with other upstream Vim improvements