Don't cleanup AppImage extraction on fuseless systems
This commit disables the self-cleanup of the AppImage on systems without
Fuse v2 (e.g. are run with `--appimage-extract-and-run`)
This was done to increase startup time of those AppImages, since the files
don't need to be rextracted on every run.

Also this avoids issues when multiple instances of AppImages are running
in parallel, because if one instance would exit, it would cleanup the
extracted files while the other is still using them.
This is caused because AppImages are extracted to the same directory
on every extraction and is not named randomly