Decrease allowed chezmoiversion to v2.37.0
This commit allows Chezmoi v2.37.0 to manage the repository.
Before that only Chezmoi v2.38.0 and above were allowed to manage the dotfiles
to ensure compatiblity of the required features of the dotfiles.

But Gentoo only ships with v2.37.0 as package so .chezmoiversion was adjusted
to also allow that version to manage the dotfiles. After a glance on the
changelog between those two versions no major breaking changes were done
so the Gentoo version should run perfectly fine