Create ~/bin by default
This commit creates ~/bin when chezmoi is run. Additionally a README
is added so the user knows for what the directory is used.

It is used for local user scripts/programs which are not managed by
the system nor chezmoi. So they won't be removed/overwritten by
external factors.
Also programs there appear first in `$PATH` so it can be used to overwrite/
mask system or chezmoi installed programs