Greatly simplify .chezmoiignore to include most configs even if not installed
This commit greatly removes complexity from the .chezmoiignore config file.
Now only systems which are headless/not headless are differentiated.

For example i3 configs are now always installed on non headless (graphical)
systems, even though i3 is not installed. This was done to make sure that
all configs are in place should i3 be installed in the future.
Also management complexity of the chezmoiignore has been dramatically
decreases