Add zoxide to chezmoi add use it in all shells
This commit adds zoxide (autojump) to the dotfiles which will allow
the user to jump quickly to visited places.

Also shell integration is implemented so all shells should be
able to use zoxide. The command is `j` just as in autojump.
Implemented shells: `bash`, `fish` and fallback `posix` which
works with `dash` and should also work with `zsh`