Split each external program into its own `.chezmoiexternals/` file
This commit splits the big headless.toml.tmpl into multiple files,
grouped by program, that is vim is now in vim.toml.tmpl, tmux in its own
and other.

This was needed because zoxide required multiple entries in chezmoiexternal
to install the binary, man pages and the completion for every shell.

Now the `.chezmoiexternals/` contains files for each program related files
to install, which makes it less confusing in what files what entry belongs to