Bump Vim AppImage to v9.0.1894
This commit updates the Vim AppImage to v9.0.1894 fixing various
AppImage startup bugs. The previous version was broken because it couldn't
source/load the default Vim runtimes files causing alot of warnings and
missing configuration and syntax.

This commit bumps the version to an AppImage version that has these
fixed