Add :tabnew to qutebrowser and its keybinding
This commit adds a commonly used Vim command to qutebrowser which was
missing.
- `:tabnew` opens a new tab in Vim and therefore this command should also
be implemented in qutebrowser to mimic Vim behavior as closely as possible.
- `<Ctrl-w><n>` also opens a new window in Vim, so I also implemented it
in qutebrowser but instead it opens a new tab.

I got the idea from Firefox Addon Tridactyl which has this implemented,
I was quite suprised that it wasn't in qutebrowser