Fix qutebrowser open download keybind unbound
This commit fixes a side-effect of the removal of the default keybinds
to make room for the custom <Ctrl-w> and <Ctrl-x>
It broke the download prompt to be able to open the directly file afterwards.
As it was bound to <Ctrl-x> but removed from every start, there was no
way to open the downloaded file directly from qutebrowser but it was needed
to open `cd` into the directory and open it manually.

This is now fixed by assigning it with a custom keybind <Shift-Return>,
that is SHIFT with ENTER, to open the file after the download automatically