Add zoxide binaries for Aarch64/arm64/armv8
This commit adds support for the Aarch64 (also referred as arm64 or
armv8) architecture for `zoxide`.

That means chezmoi will now install on Aarch64 platform the zoxide
binary with manpages and shell completion.
This is a preparation for having a full fledged system on Aarch64
systems like Android or Raspberry Pi