Add chezmoiignore description and rules for ignoring externals
This commit improves the current description for `.chezmoiignore`
inparticular notes about where and how to exclude external binaries
added by `.chezmoiexternal*` files/entries.

External files should not be ignored by chezmoiignore but rather by the
external entry itself by making it blank if the binary/external file should
not be installed.

This should help in maintaining exclusion rules for files which are
not directly in the dotfiles repository and thus making the ignore file
less messy