Add glibc detection in chezmoi config file
This commit adds a new template variable `.glibc` which results to
either `true` or `false` if glibc (the standard GNU C library) is
installed.
This is required as most AppImages are compiled against glibc so they
require a `glibc` system to run on. To filter these AppImages out of non
`glibc` systems (like Alpine Linux which runs `musl`) this variable can
be used for flow control to exclude or include the binaries based on
`glibc` compatibility

Also this commit fixes the battery and wireless device detection which
was broken caused by invalid string comparison