Only install AppImages on Glibc systems
This commit makes sure that the current AppImages (vim and tmux) are
only installed by chezmoi on systems which have the glibc standard C
library and not others like musl.

The Glibc is required for AppImages as they have a dependency on it and
cannot run with other standard C libraries like musl.

Now they won't be installed on those systems preventing AppImage crashes
and trouble for the user