Add shebang #! referencing template for fixing broken shebangs
This commit adds the shebang.tmpl template that helps settings the
correct shebang (#!) for scripts on systems which do not have default
path like /usr/bin/env
It will look for the given path, if it doesn't exist it will replace it
with a working one.

This is required for some non compliant/weird systems like Termux on Android.
Example template usage:

    #! {{ template "shebang.tmpl" "/usr/bin/env" }} python

The template will make sure that /usr/bin/env exist and is executable,
if not try to find the binary from $PATH and use that path instead.
For example on Termux /usr/bin/env does not exist, it only exist at
/data/data/com.termux/files/usr/bin/env so the script execution would
fail. In the Termux case, using the template would replace /usr/bin/env
with /data/data/com.termux/files/usr/bin/env which would result to the
following shebang line:

    #! /data/data/com.termux/files/usr/bin/env python

Obviously one could use the template directly with the required binary
like /usr/bin/python and if not found the template would resolve it via
the $PATH just like env would. But for non-chezmoi scripts /usr/bin/env
should be used. See below for detailed description.

The /usr/bin/env sheband path should be used to find an run the correct
interpreter for the given script as it makes the script more portable,
as not all machines have the same path for the same binary.
/usr/bin/env searches $PATH for the binary and executes the first match.

For example python on machine A might be installed at /usr/bin/python
but on machine B at /lib64/bin/python; using a fixed path shebang
would break the execution of it when run on different machines.
Instead one should use `/usr/bin/env python`, env will then try to find
the python binary python in the $PATH env variable.

/usr/bin/env is not a standard path but pretty much used in every system
(except Termux caused by technical restrictions)