Make use of shebang template for run_after scripts
This commit changes the shebang for the scripts that chezmoi runs after
it has applied the dotfiles.
The changed shebangs use the shebang.tmpl template to ensure working
shebangs even on systems which do not have the given path for the
binary. Instead it will search for the binary and use the fixed shebang
instead.

One such system with weird path for env is Termux on Android