Print instructions for shell reload after chezmoi apply
This commit adds a shell script which will run after chezmoi has
successfully applied the source state to the target state.

It will print instructions on how to reload/restart the current shell to
reload the new configuration applied with chezmoi. Reloading the shell
is in most cases the last step to do after dotfiles installation/update.

This is useful on systems where the current shell is not standard or is
unknown to the user. This will give the user easy steps to follow for
reloading a possibly new and fresh shell configuration