Suppress warnings for AppImages on fuseless systems
This commit changes the mechanism how AppImages are executed on systems
which do not have fuse v2 installed.
Instead of calling `--appimage-extract-and-run` we extract the AppImage
manually with `--appimage-extract` to a isolated directory which
only the user can access (for additional security)

Since we extract the AppImage ourself, annoying warnings like
`File exists and file size matches, skipping` can be avoided as we
only extract the AppImage if not already done.

The temporary directory will be created at /tmp/$USER.AppImages/ which
will contain all user extracted AppImages, allowing multiple
users to run the same AppImage in parallel as the file owner is always
the same user