Add fzf binary and its shell completions to chezmoi
This commit brings the full power of fzf to bash and fish.
That means that fzf will now be installed by chezmoi (for arm64 and amd64)
including keybindings for bash and fish shell to efficiently use the cli
with fzf.

Custom keybindings:
- Ctrl-T: fuzzy find file to add to cli
- Ctrl-R: fuzzy find through shell history
- Alt-C:  fuzzy find path to cd into

Also in this commit the loading of custom bash configs from ~/.bashrc.d/
was added, making it possible to just drop the *.bash files into the
directory, which bash will load automatically (similiar to fish)