Add bash prompt git status indicator
This commit extends the bash prompt by showing the current git status
flags and the current branch next to the current path in the bash prompt.

This functionality is from a bash script from the git repo which will be
downloaded and installed to make sure the functionality is available on
all systems.

Additionally special git state flags are shown on bash prompt:
- * unstaged files
- + staged files
- $ stashed files
- % untracked files

In directories without .git directory (not git tracked) the prompt is hidden