Install Linux binaries only on supported platforms
This commit makes sure that only on fully-fledged Linux systems
Linux binaries are installed. This mostly affects Android,
as Linux binaries seem to be compatible with Android but on it some
required syscalls are blocked (especially faccessat2), so these Linux
binaries would crash.

Instead on Android specially compiled Android binaries should be used
as these workaround these limitations.

For example for Go Android binaries should be compiled with
GOOS=android and the default GOOS=linux binaries should not be used.

This also might affect other languages (for example rust or C) but this
couldn't be verified.

Also this commit makes sure on Windows/OS X the Linux binaries are also
not installed