Inform user if important cli binaries are missing
This commit adds a run_after_ chezmoi script that verifies that
important binaries like fzf, tmux or vim are installed.

If those are not present their config files for example bash completion
or shell keybindings are installed nevertheless and the user is informed
that the binaries should be installed.

This is particular useful for systems where the binaries are not available
for example on Android fzf, zoxide and other statically build binaries are
not installed by chezmoi due to ABI incompatibility. Instead make sure the
user knows that packages have to be installed afterwards