Fix invalid awk arguments on non GNU linux systems
This commit fixes issues related to awk GNU extensions not available on
BusyBox implementation. Some parameters were not available in BusyBox,
instead use ones available on BusyBox awk and GNU awk.

Awk is used to extract the current user shell from /etc/passwd to display
the correct command for reloading the shell.

Also this commit fixes that AppImages vim and tmux are downloaded and
installed on non Glibc systems, which is wrong as they cannot be executed,
as AppImages require Glibc to be installed