Add screenshot wrapper script as keybind to i3
This commit adds a screenshot wrapper script that finds an installed
screenshot utility and takes a screenshot with it, regardless the
installed screenshot tools.
The user doesn't need to remember the systems screenshot tool and its
parameter, instead the user can simply run this script and it will find
and use the correct parameter to create a screenshot.

The i3 keybind is <Win+Print> and <Win+Shift+Print> (in passthrough mode)
to create a screenshot with the added script.
By default the user will then need to select a rectangular area that will
be screenshotted.

All screenshots created by the script will be saved into
~/Pictures/screenshot-$(date).png and to the clipboard