Theme (selenized) dunst notification daemon and set its font
This commit improves the visual appearance of notifications displayed
with dunst. They now use the same font as i3 widgets and they are now
colored with selenized color palette.

This aligns the visual appereance better to the overall system.

Also the notification of the screenshot utility has been improved by
including also a title in the notification