Add external Go program `gojq` as portable `jq` implementation
This commit adds an jq implementation written in Go, called `gojq`.
Since it is statically compiled this allows distribution via chezmoi.

It was mainly added for future use with i3-msg, which prints its output
as JSON. In combination those can be used for complex i3 scripting in
shell scripts.

For example for renaming the current workspace without requiring to
enter the current workspace number requires reading JSON output,
the added binary should make this possible