Remove .vimrc on chezmoi run
This commit removes the ~/.vimrc file on chezmoi apply.
This is required as when using the vim AppImage, the defaults.vim is not
sourced when the ~/.vimrc is exists. But the defaults.vim is important
to be loaded as it set filetype, syntax, indentation and other good vim
stuff, which won't be loaded when a ~/.vimrc exists.

I could add the line to ~/.vimrc that it should source it, but I want to
do that in the future when I get the time to setup vim to its best.
For now rely on defaults, which are good enough.
Without defaults.vim Vim is ugly and close to useless as most important
features are missing.

Before applying the update to the target state (to the machine), one
should backup the current vimrc to something like ~/.vimrc.old, so in
the future I can look at these settings and pick the best and add those
to the holy vimrc when I get time to do that