Add i3 workspace rename script with persistent workspace number
This commit adds a script that allows renaming of the current or a
specified workspace without changing the current workspace number/id
except when its explicitly given in the new name.

This allows renaming of the workspace without breaking the shortcuts for it.
For example the workspace '3: test' should be renamed to 'mail',
the result would be '3: mail'.  The shortcuts still work as the workspace
number hasn't been changed.

Additionally it allows moving of workspaces by giving a workspace number
in the name explicitly
For example: '3: test' renamed to '6: mail' will work as expected