Improve i3 workspace buttons theme
This commit improves the visibility of the workspace buttons to increase
distinction between them. This was needed because with i3 workspace
renaming they where hard to distinguish as the buttons had different
length.

Now the background and the border colors have a greater contrast and
make it easier to distinguish them