Remove unused dependency from run_after_check_cli_binaries_init.sh
This commit removes `perl` from the dependency checking script, as it is
not needed anymore.
The only binary that had a dependency on it was fzf, but on the most
recent release perl is not required anymore as fzf will resort to awk or
mawk instead if it cannot find it.

As awk is a standard system utility it should be used instead, freeing
the dependency on perl, which is useful for bare-bones systems like
Alpine which do not have perl by default installed