Add ble.sh bash commandline editor
This commit adds the ble.sh `Bash Line Editor` to bash.
This enables advanced autocompletion and syntax highlighting in the cli.

Very similar to fishs behavior and useful on systems which do not have
fish installed

Need to test its capabilities and stability further, in the future I
might downgrade to a more stable version (this is the devel version) or
I might get rid of ble.sh altogether if I don't see it fit with my setup