Fix chezmoi init fails on WSL Ubuntu due to uncommon /sys NIC structure
This commit fixes issues in the .chezmoi.toml.tmpl making chezmoi
crash/abort while in the init phase.

The issue is caused because on WSL the NIC files in /sys have a
different layout then the usual Linux system. The current config expects
a directory, but under WSL those are files. The check is needed to
determine if the current machine has a wireless devices, this will
decide if wireless related configs should also be installed onto the
system.

This commit makes sure to check only for wireless devices when the /sys
layout contains directories and not a file