Remove chezmoi unmanaged files from ~/.bashrc.d and ~/.local/share/blesh
This commit ensures that when the version of ble.sh is updated or
contents from .bashrc.d has been changed, those changes are also
propagated onto other machines ensuring that leftover files, which have
been renamed or deleted upstream are also deleted from the machine.

This is done by adding the exact_ attribute to chezmoi managed
directories. Chezmoi will then remove files which are not found in the
source state from the exact_ target state.