Make i3 startable via startx
This commit create the .xinitrc file which is being executed by startx.
This allows startx to run i3 from the commandline.

Also the dbus-session is started when being run via .xinitrc or .xsession,
it is required for various Flatpak apps, like Chromium which won't even start
without a running dbus session. (hence .xsession and .xinitrc are
executed from a CLI environment, the DBus session for the current user
is not running, so the manual run is needed.