Merge pwrshutdown and pwrsuspend scripts into one big wrapper
This merges the pwrshutdown and pwrsuspend scripts into one big wrapper
which does now all power controlling.
Additionally for every desirable power state a symlink was created in
~/.local/bin which points to the newly created pwrctrl script,
it will handle the desired state based on the symlink with which it was
called with.

For example: pwrhibernate -> pwrctrl
pwrctrl knows its been called as pwrhibernate therefore it'll hibernate
the system.

Following symlinks/states are available:
- pwrhibernate (hibernate)
- pwrhybrid (suspend and hibernate)
- pwrreboot (reboot)
- pwrshutdown (shutdown)
- pwrsuspend (suspend)
- pwrstaged (suspend first then after timeout hibernate)

Note: the rootless nature of the previous wrappers are kept,
that is the script tries to do all actions without root rights,
if possible.