Fix remote i3wm session creation for systems without i3 installed
This commit addresses a issue where systems lacking the i3 window
manager may experience problems when trying to create a remote desktop
connection. This situation occurs on non-headless systems running an
alternative (non i3) window manager or desktop environment.

Previously, connecting via RDP to the Xrdp service would result in a
connection loss due to the session being set to execute i3 in both the
`~/.xinitrc` and `~/.xsession` files.

This fix ensures not to set i3 as WM when its not installed for remote
desktops