Bump minimum Chezmoi version to 2.43.0
This commit sets the minimum required chezmoi version for use with the
repository to 2.43.0

This is required as version 2.43.0 fixes an issue with `exact_` folders
and .chezmoiexternals pointing to that directory. Previously the issue
led to removal of the externals from that `exact_` folder, thus leaving
the target state in a wrong state.
To prevent this issue, version 2.43.0 must be used.

PS: the issue was reported by me :)