Add Anti-CAPS-Stuck keybinding
This commit adds a CAPS triggering mechanism should the keyboard be
changed (to the one with CAPS remapped to ESC).
When triggering this with Ctrl-Shift-Caps CAPS will be virtually
pressed, releasing Caps mode if we were in it.
This helps in unstucking in Caps because when the Caps remapping is
applied but we're in Caps mode, we would'nt break out of it