Move rofi single line prompt into an extra executable (rofi-prompt)
Extract rofis single line prompt into its own executable rofi-prompt for
use in other scripts. Extracted out of i3-workspace-rename

It was planned to use this script in i3-mark and i3-jump, but as it
stands currently its not useable because rofi cannot auto-accept input
after X amount of characters rendering this solution useless to me.

Nevertheless, this change could prove useful in the future