Revert Selenized colors to match reference values
The Selenized reference `the-value.md` contains incorrect
values for the `bg_1` and `bg_2` colors in the light and dark variants.
Although the previous commit corrected these values,
this commit reverts them to match the incorrect reference values.

This ensures consistency, as the incorrect reference is still used widely.
Until the reference is officially updated, we'll be using the incorrect
values.

For more details, see: https://github.com/jan-warchol/selenized/issues/99