Fix rclone being falsely deployed on Android OS
This commit prevents `rclone` from being deployed onto Android as the
default Linux binary is not really compatible on Android systems and
cannot be used as DNS resolution is not correctly working, etc.

Instead a binary specifically compiled for the Android OS must be used.
Currently `rclone` upstream does not provide the required binary
therefore it must be installed manually via the package manager of
choice.