Add recommended cli binary check
This commit adds a check for potentially missing cli binaries which are
not required but recommended to be installed.
Previously inly required binaries are checked but now also optional
(nice-to-have) binaries are being checked for and a info is printed
instead of a warning.
The first recommended binary checked is `rclone`