Fix ble.sh cache directory being managed by Chezmoi
This commit prevents Chezmoi to inspect/manage ble.sh dynamic/cache
directory which ble.sh uses internally to cache/store things.
Since it content is dynamic and always changing, Chezmoi should not
manage it as on every `chezmoi apply` it warns of changed files.
The fix adds an ignore entry to `.chezmoiignore.tmpl`