Enforce .local/bin to replicate exact Chezmoi source state
This commit makes the `.local/bin` to be handled by Chezmoi with the
`exact: true` property.
This causes all files in the directory not available in the Chezmoi
source state to be removed by Chezmoi.
Thus this cleans the folder from residual waste.

For user added binaries, not managed by Chezmoi, use the `~/bin`
directory