Add container bash prompting
When inside a container bash PS1 prompt has now the container name added
to it.
This makes it easy to distinguish whether the user is inside a container or
not.
This is especially useful when inside distrobox/toolbox where different
containers have different purposes, now its possible to check if inside
the correct container